pyproject.toml
[project]
name = "sql-assistant-with-local-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==5.3.0",
    "attrs==23.2.0",
    "blinker==1.8.2",
    "cachetools==5.3.3",
    "certifi==2024.7.4",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "ctransformers==0.2.27",
    "filelock==3.15.4",
    "fsspec==2024.6.1",
    "gitdb==4.0.11",
    "gitpython==3.1.43",
    "huggingface-hub==0.23.4",
    "idna==3.7",
    "jinja2==3.1.4",
    "jsonschema==4.22.0",
    "jsonschema-specifications==2023.12.1",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "networkx==3.3",
    "numpy==2.0.0",
    "packaging==24.1",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "protobuf==5.27.2",
    "py-cpuinfo==9.0.0",
    "pyarrow==16.1.0",
    "pydeck==0.9.1",
    "pygments==2.18.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rich==13.7.1",
    "rpds-py==0.18.1",
    "six==1.16.0",
    "smmap==5.0.1",
    "streamlit==1.36.0",
    "sympy==1.12.1",
    "tenacity==8.5.0",
    "toml==0.10.2",
    "toolz==0.12.1",
    "torch==2.3.1",
    "torchaudio==2.3.1",
    "torchvision==0.18.1",
    "tornado==6.4.1",
    "tqdm==4.66.4",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "urllib3==2.2.2",
    "watchdog==4.0.1",
]
